+++ KD_Lib/__init__.py
@@ -7,6 +7,6 @@ __email__ = 'divhet163@gmail.com'
 __version__ = '0.0.1'
 
 from .models import resnet
-from .original import mnist, original
+from .original import original
 from .attention import attention
 from .TAKD import TAKD
