+++ KD_Lib/original/__init__.py
@@ -1,2 +1,1 @@
-from . import mnist, train, evaluate, model
 from KD_Lib.original.original_paper import original
