+++ KD_Lib/original/original_paper.py
@@ -1,18 +1,19 @@
-import torch
 import torch.nn as nn
 import torch.nn.functional as F
-import torch.optim as optim
-from .model import teacher, student
 from KD_Lib.RKD import RKDLoss
 from KD_Lib.common import BaseClass
 
+
 class original(BaseClass):
-    def __init__(self, teacher_model, student_model, train_loader, val_loader, optimizer_teacher, optimizer_student, loss='MSE', temp=20.0, distil_weight=0.5, device='cpu', rkd_angle=None, rkd_dist=None):
+    def __init__(self, teacher_model, student_model, train_loader, val_loader,
+                 optimizer_teacher, optimizer_student, loss='MSE', temp=20.0,
+                 distil_weight=0.5, device='cpu', rkd_angle=None,
+                 rkd_dist=None):
         super(original, self).__init__(
-            teacher_model, 
+            teacher_model,
             student_model,
             train_loader,
-            val_loader, 
+            val_loader,
             optimizer_teacher,
             optimizer_student,
             loss,
@@ -27,13 +28,14 @@ class original(BaseClass):
             self.loss_fn = nn.KLDivLoss()
 
         elif self.loss.upper() == 'RKD':
-            self.loss_fn = RKDLoss(dist_ratio=rkd_dist, angle_ratio=rkd_angle)  
+            self.loss_fn = RKDLoss(dist_ratio=rkd_dist, angle_ratio=rkd_angle)
 
     def calculate_kd_loss(self, y_pred_student, y_pred_teacher, y_true):
         soft_teacher_out = F.softmax(y_pred_teacher/self.temp, dim=0)
         soft_student_out = F.softmax(y_pred_student/self.temp, dim=0)
 
-        loss = (1-self.distl_weight) * F.cross_entropy(soft_student_out, y_true)
-        loss += self.distl_weight * self.loss_fn(soft_teacher_out, soft_student_out)
-
-        return loss 
+        loss = (1-self.distl_weight) * F.cross_entropy(soft_student_out,
+                                                       y_true)
+        loss += self.distl_weight * self.loss_fn(soft_teacher_out,
+                                                 soft_student_out)
+        return loss
